+++ src/components/BorrowerProfile/TrusteeDetails.vue
@@ -10,35 +10,56 @@
 		<dl class="tw-mb-4">
 			<description-list-item
 				:term="'Kiva borrowers'"
-				:details="numLoansEndorsedPublic"
-				@show-definition="$emit('show-definition',
-					{cid: 'bp-def-trustee-kiva-borrowers', sfid: '50150000000s2cE'})"
+				:details="numLoansEndorsedPublicFormatted"
+				@show-definition="$emit('show-definition', {
+					cid: 'bp-def-trustee-kiva-borrowers',
+					sfid: '50150000000s2cE',
+					panelName: 'Trustee',
+					linkText: 'Kiva borrowers'
+				})"
 			/>
 			<description-list-item
 				:term="'Total loans'"
 				:details="totalLoansValueFormatted"
-				@show-definition="$emit('show-definition',
-					{cid: 'bp-def-trustee-total-loans', sfid: '50150000000s2cO'})"
+				@show-definition="$emit('show-definition', {
+					cid: 'bp-def-trustee-total-loans',
+					sfid: '50150000000s2cO',
+					panelName: 'Trustee',
+					linkText: 'Total loans'
+				})"
 			/>
 			<description-list-item
 				:term="'Loans defaulted'"
-				:details="numDefaultedLoans"
-				@show-definition="$emit('show-definition',
-					{cid: 'bp-def-trustee-loans-defaulted', sfid: '50150000000s2cn'})"
+				:details="numDefaultedLoansFormatted"
+				@show-definition="$emit('show-definition', {
+					cid: 'bp-def-trustee-loans-defaulted',
+					sfid: '50150000000s2cn',
+					panelName: 'Trustee',
+					linkText: 'Loans defaulted'
+				})"
 			/>
 			<description-list-item
 				:term="'Repayment rate'"
 				:details="repaymentRateFormatted"
-				@show-definition="$emit('show-definition',
-					{cid: 'bp-def-trustee-repayment-rate', sfid: '50150000000s2cx'})"
+				@show-definition="$emit('show-definition', {
+					cid: 'bp-def-trustee-repayment-rate',
+					sfid: '50150000000s2cx',
+					panelName: 'Trustee',
+					linkText: 'Repayment rate'
+				})"
 			/>
 		</dl>
 		<kv-text-link
 			:icon="mdiArrowRight"
 			:href="`/trustees/${trusteeId}`"
 			target="_blank"
+			v-kv-track-event="[
+				'Borrower profile',
+				'click-Trustee-tab',
+				noTrusteeState ? 'More about loans without endorsements' : 'More about this trustee'
+			]"
 		>
-			More about this trustee
+			{{ noTrusteeState ? 'More about loans without endorsements' : 'More about this trustee' }}
 		</kv-text-link>
 		<div v-if="endorsement" class="tw-prose tw-my-3">
 			<h3>{{ endorsementTitle }}</h3>
@@ -63,6 +84,10 @@ export default {
 		KvTextLink,
 	},
 	props: {
+		borrowerName: {
+			type: String,
+			default: '',
+		},
 		endorsement: { // endorsement
 			type: String,
 			default: '',
@@ -99,7 +124,10 @@ export default {
 	},
 	computed: {
 		endorsementTitle() {
-			return `Why are you endorsing ${this.trusteeName}?`;
+			return `Why are you endorsing ${this.borrowerName}?`;
+		},
+		noTrusteeState() {
+			return this.trusteeName === 'No Trustee Endorsement';
 		},
 		totalLoansValueFormatted() {
 			return numeral(this.totalLoansValue).format('$0,0[.]00');
@@ -107,6 +135,12 @@ export default {
 		repaymentRateFormatted() {
 			return numeral(this.repaymentRate / 100).format('0%');
 		},
+		numLoansEndorsedPublicFormatted() {
+			return numeral(this.numLoansEndorsedPublic).format('0,0');
+		},
+		numDefaultedLoansFormatted() {
+			return numeral(this.numDefaultedLoans).format('0,0');
+		}
 	},
 };